+++ packages/editor-common/web/src/consts.js
@@ -25,6 +25,7 @@ export const COMMANDS = Object.freeze({
   DELETE: 'delete',
   TAB: 'tab',
   SHIFT_TAB: 'shiftTab',
+  ESC: 'esc',
 });
 
 export const TEXT_TYPES = Object.freeze([