+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - [#1241](https://github.com/wix-incubator/rich-content/pull/1241) insert link to image will not close the toolbar
 - `editor`
   - [#1243](https://github.com/wix-incubator/rich-content/pull/1243) when pasting text with hyperlinks, spaces deleted
+  - [#1252](https://github.com/wix-incubator/rich-content/pull/1252) blur editor on esc keypress
 
 ### :house: Internal
 - `general`